initial changes committed
