+++ http/src/main/scala/org/broadinstitute/dsde/workbench/leonardo/http/service/AppServiceInterp.scala
@@ -10,7 +10,6 @@ import cats.effect.std.Queue
 import cats.mtl.Ask
 import cats.syntax.all._
 import org.apache.commons.lang3.RandomStringUtils
-import org.broadinstitute.dsde.workbench.azure.{AKSClusterName, RelayNamespace}
 import org.broadinstitute.dsde.workbench.google2.GKEModels.{KubernetesClusterName, NodepoolName}
 import org.broadinstitute.dsde.workbench.google2.KubernetesSerializableName.NamespaceName
 import org.broadinstitute.dsde.workbench.google2.{
@@ -19,8 +18,6 @@ import org.broadinstitute.dsde.workbench.google2.{
   GoogleResourceService,
   KubernetesName,
   MachineTypeName,
-  NetworkName,
-  SubnetworkName,
   ZoneName
 }
 import org.broadinstitute.dsde.workbench.leonardo.AppRestore.{CromwellRestore, GalaxyRestore}
@@ -28,13 +25,7 @@ import org.broadinstitute.dsde.workbench.leonardo.AppType._
 import org.broadinstitute.dsde.workbench.leonardo.JsonCodec._
 import org.broadinstitute.dsde.workbench.leonardo.SamResourceId._
 import org.broadinstitute.dsde.workbench.leonardo.config._
-import org.broadinstitute.dsde.workbench.leonardo.dao.LandingZoneResourcePurpose.{
-  AKS_NODE_POOL_SUBNET,
-  LandingZoneResourcePurpose,
-  SHARED_RESOURCE,
-  WORKSPACE_BATCH_SUBNET
-}
-import org.broadinstitute.dsde.workbench.leonardo.dao.{LandingZoneResourcesByPurpose, WsmDao}
+import org.broadinstitute.dsde.workbench.leonardo.dao.WsmDao
 import org.broadinstitute.dsde.workbench.leonardo.db.KubernetesServiceDbQueries.getActiveFullAppByWorkspaceIdAndAppName
 import org.broadinstitute.dsde.workbench.leonardo.db._
 import org.broadinstitute.dsde.workbench.leonardo.http.service.LeoAppServiceInterp.isPatchVersionDifference
@@ -42,12 +33,10 @@ import org.broadinstitute.dsde.workbench.leonardo.model.SamResourceAction._
 import org.broadinstitute.dsde.workbench.leonardo.model._
 import org.broadinstitute.dsde.workbench.leonardo.monitor.LeoPubsubMessage._
 import org.broadinstitute.dsde.workbench.leonardo.monitor.{ClusterNodepoolAction, LeoPubsubMessage}
-import org.broadinstitute.dsde.workbench.leonardo.util.AppCreationException
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 import org.broadinstitute.dsde.workbench.model.{TraceId, UserInfo, WorkbenchEmail}
 import org.broadinstitute.dsde.workbench.openTelemetry.OpenTelemetryMetrics
 import org.broadinstitute.dsp.{ChartVersion, Release}
-import org.http4s.headers.Authorization
 import org.http4s.{AuthScheme, Uri}
 import org.typelevel.log4cats.StructuredLogger
 
@@ -501,7 +490,7 @@ final class LeoAppServiceInterp[F[_]: Parallel](config: AppServiceConfig,
         case CloudProvider.Gcp => F.pure(None)
         case CloudProvider.Azure =>
           for {
-            landingZoneResources <- getLandingZoneResources(workspaceDesc.spendProfile, userToken)
+            landingZoneResources <- wsmDao.getLandingZoneResources(workspaceDesc.spendProfile, userToken)
           } yield Some(landingZoneResources)
       }
 
@@ -682,78 +671,6 @@ final class LeoAppServiceInterp[F[_]: Parallel](config: AppServiceConfig,
         } yield ()
       }
   } yield ()
-
-  private def getLandingZoneResources(billingProfileId: String, userToken: Authorization)(implicit
-    ev: Ask[F, AppContext]
-  ): F[LandingZoneResources] =
-    for {
-      // Step 1: call LZ for LZ id
-      landingZoneOpt <- wsmDao.getLandingZone(billingProfileId, userToken)
-      landingZone <- F.fromOption(
-        landingZoneOpt,
-        AppCreationException(s"Landing zone not found for billing profile ${billingProfileId}")
-      )
-      landingZoneId = landingZone.landingZoneId
-
-      // Step 2: call LZ for LZ resources
-      lzResourcesByPurpose <- wsmDao.listLandingZoneResourcesByType(landingZoneId, userToken)
-
-      aksClusterName <- getLandingZoneResourceName(lzResourcesByPurpose,
-                                                   "Microsoft.ContainerService/managedClusters",
-                                                   SHARED_RESOURCE,
-                                                   false
-      )
-      batchAccountName <- getLandingZoneResourceName(lzResourcesByPurpose,
-                                                     "Microsoft.Batch/batchAccounts",
-                                                     SHARED_RESOURCE,
-                                                     false
-      )
-      relayNamespace <- getLandingZoneResourceName(lzResourcesByPurpose,
-                                                   "Microsoft.Relay/namespaces",
-                                                   SHARED_RESOURCE,
-                                                   false
-      )
-      storageAccountName <- getLandingZoneResourceName(lzResourcesByPurpose,
-                                                       "Microsoft.Storage/storageAccounts",
-                                                       SHARED_RESOURCE,
-                                                       false
-      )
-      vnetName <- getLandingZoneResourceName(lzResourcesByPurpose, "DeployedSubnet", AKS_NODE_POOL_SUBNET, true)
-      batchNodesSubnetName <- getLandingZoneResourceName(lzResourcesByPurpose,
-                                                         "DeployedSubnet",
-                                                         WORKSPACE_BATCH_SUBNET,
-                                                         true
-      )
-      aksSubnetName <- getLandingZoneResourceName(lzResourcesByPurpose, "DeployedSubnet", AKS_NODE_POOL_SUBNET, true)
-    } yield LandingZoneResources(
-      AKSClusterName(aksClusterName),
-      BatchAccountName(batchAccountName),
-      RelayNamespace(relayNamespace),
-      StorageAccountName(storageAccountName),
-      NetworkName(vnetName),
-      SubnetworkName(batchNodesSubnetName),
-      SubnetworkName(aksSubnetName)
-    )
-
-  private def getLandingZoneResourceName(landingZoneResourcesByPurpose: List[LandingZoneResourcesByPurpose],
-                                         resourceType: String,
-                                         purpose: LandingZoneResourcePurpose,
-                                         useParent: Boolean
-  ): F[String] =
-    landingZoneResourcesByPurpose
-      .filter(_.purpose == purpose)
-      .flatMap(_.deployedResources)
-      .filter(_.resourceType.equalsIgnoreCase(resourceType))
-      .headOption
-      .flatMap { r =>
-        if (useParent) r.resourceParentId.flatMap(_.split('/').lastOption)
-        else r.resourceName.orElse(r.resourceId.flatMap(_.split('/').lastOption))
-      }
-      .fold(
-        F.raiseError[String](
-          AppCreationException(s"${resourceType} resource with purpose ${purpose} not found in landing zone")
-        )
-      )(F.pure)
 
   private[service] def getSavableCluster(
     userEmail: WorkbenchEmail,
