+++ http/src/main/scala/org/broadinstitute/dsde/workbench/leonardo/util/AzurePubsubHandlerAlgebra.scala
@@ -64,8 +64,8 @@ trait AzurePubsubHandlerAlgebra[F[_]] {
 
 final case class CreateAzureRuntimeParams(workspaceId: WorkspaceId,
                                           runtime: Runtime,
-                                          relayeNamespace: RelayNamespace,
                                           storageContainerResourceId: WsmControlledResourceId,
+                                          landingZoneResources: LandingZoneResources,
                                           runtimeConfig: RuntimeConfig.AzureConfig,
                                           vmImage: AzureImage
 )
